+++ include/ambulant/gui/gtk/gtk_video_renderer.h
@@ -53,7 +53,7 @@ class gtk_video_renderer :  public gtk_renderer<common::video_renderer>  {
 	net::pixel_order pixel_layout();
 	void redraw_body(const lib::rect &r, common::gui_window* w);
   protected:
-	void _push_frame(char* frame, int size);
+	void _push_frame(char* frame, size_t size);
 private:
 	std::queue< std::pair<int, char*> > m_frames;
 	long int m_img_displayed;
