+++ include/ambulant/net/stdio_datasource.h
@@ -65,14 +65,14 @@ class stdio_datasource : virtual public datasource, virtual public lib::ref_coun
 	void start_prefetch(ambulant::lib::event_processor *evp){};
 #endif
 	void stop();
-	void readdone(int len);
+	void readdone(size_t len);
 	bool end_of_file();
 
 
 	char* get_read_ptr();
-	int size() const;
+	size_t size() const;
 
-	void read(char *data, int size);
+	void read(char *data, size_t size);
 
 	friend inline std::ostream& operator<<(std::ostream& os, const stdio_datasource& n) {
 		os << "stdio_datasource(" << (void *)&n << ", source=" << n.m_url.get_url() << ")";
