+++ include/ambulant/net/win32_datasource.h
@@ -63,11 +63,11 @@ class win32_datasource : virtual public datasource, virtual public lib::ref_coun
 
 	void start(ambulant::lib::event_processor *evp, ambulant::lib::event *callback);
 	void stop();
-	void readdone(int len);
+	void readdone(size_t len);
 	bool end_of_file();
 	char* get_read_ptr();
-	int size() const;
-	void read(char *data, int size);
+	size_t size() const;
+	void read(char *data, size_t size);
   #ifdef WITH_SEAMLESS_PLAYBACK
 	void start_prefetch(ambulant::lib::event_processor *evp){};
 #endif
